backend/requirements.txt
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.12.14
cffi==1.17.1

celery[redis]==5.4.0

channels==4.2.0
channels_redis==4.2.1

charset-normalizer==3.4.1
constantly==23.10.4
cryptography==44.0.0

daphne==4.1.2
defusedxml==0.7.1
Django==5.1.5
django-cors-headers==4.6.0
django-extensions==3.2.3
django-debug-toolbar==4.4.6

djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0

h2==4.1.0
hpack==4.1.0
httpie==3.2.4
hyperframe==6.1.0
hyperlink==21.0.0

idna==3.10
incremental==24.7.2
iniconfig==2.0.0

markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0

packaging==24.2
pluggy==1.5.0
priority==1.3.0

psycopg2-binary==2.9.9

pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
PySocks==1.7.1

pytest==8.3.4
pytest-django==4.9.0

redis==5.2.1

requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4

service-identity==24.2.0
setuptools==75.8.0

sqlparse==0.5.3

Twisted==24.11.0
txaio==23.1.1

typing_extensions==4.12.2
urllib3==2.3.0
zope.interface==7.2